@@ -4,55 +4,54 @@ import (
 	"context"
 	"fmt"
 	"math/rand"
-	"net/url"
-	"os"
-	"os/signal"
-	"path/filepath"
-	"strings"
-	"syscall"
 	"time"
 
-	nc "github.com/geniusrabbit/notificationcenter"
-	"github.com/jinzhu/gorm"
-	ugrpc "github.com/sspserver/udetect/transport/grpc"
-	uhttp "github.com/sspserver/udetect/transport/http"
+	"github.com/fasthttp/router"
+	nc "github.com/geniusrabbit/notificationcenter/v2"
+	"github.com/sspserver/udetect/transport/http"
 	"github.com/valyala/fasthttp"
 	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
+
+	"geniusrabbit.dev/adcorelib/admodels/types"
+	"geniusrabbit.dev/adcorelib/adsource"
+	"geniusrabbit.dev/adcorelib/adsource/openrtb"
+	"geniusrabbit.dev/adcorelib/eventtraking/eventgenerator"
+	"geniusrabbit.dev/adcorelib/eventtraking/eventstream"
+	"geniusrabbit.dev/adcorelib/eventtraking/pixelgenerator"
+	"geniusrabbit.dev/adcorelib/httpserver"
+	"geniusrabbit.dev/adcorelib/httpserver/extensions/endpoint"
+	"geniusrabbit.dev/adcorelib/httpserver/extensions/pixel"
+	"geniusrabbit.dev/adcorelib/httpserver/extensions/trakeraction"
+	"geniusrabbit.dev/adcorelib/httpserver/wrappers/httphandler"
+	"geniusrabbit.dev/adcorelib/net/fasthttp/middleware"
+	"geniusrabbit.dev/adcorelib/personification"
+	"geniusrabbit.dev/adcorelib/simplepersondetector"
+	"geniusrabbit.dev/adcorelib/storage/accessors/adsourceaccessor"
+	"geniusrabbit.dev/adcorelib/storage/accessors/companyaccessor"
+	"geniusrabbit.dev/adcorelib/storage/accessors/formataccessor"
+	"geniusrabbit.dev/adcorelib/storage/accessors/zoneaccessor"
+	"geniusrabbit.dev/adcorelib/urlgenerator"
+	"geniusrabbit.dev/adcorelib/zlogger"
 
 	"geniusrabbit.dev/sspserver/cmd/sspserver/appcontext"
-	"geniusrabbit.dev/sspserver/cmd/sspserver/httpserver"
-	"geniusrabbit.dev/sspserver/internal/billing"
-	"geniusrabbit.dev/sspserver/internal/data/adsourceaccessor"
-	"geniusrabbit.dev/sspserver/internal/data/loaders"
-	"geniusrabbit.dev/sspserver/internal/data/targetaccessor"
-	"geniusrabbit.dev/sspserver/internal/endpoint"
-	_ "geniusrabbit.dev/sspserver/internal/endpoint/init"
-	"geniusrabbit.dev/sspserver/internal/eventtraking/eventgenerator"
-	"geniusrabbit.dev/sspserver/internal/eventtraking/eventstream"
-	"geniusrabbit.dev/sspserver/internal/eventtraking/pixelgenerator"
-	"geniusrabbit.dev/sspserver/internal/middleware"
-	"geniusrabbit.dev/sspserver/internal/models"
-	"geniusrabbit.dev/sspserver/internal/models/types"
-	"geniusrabbit.dev/sspserver/internal/personification"
-	"geniusrabbit.dev/sspserver/internal/ssp"
-	_ "geniusrabbit.dev/sspserver/internal/ssp/platform/init"
+	"geniusrabbit.dev/sspserver/cmd/sspserver/datainit"
+	"geniusrabbit.dev/sspserver/internal/endpoint/direct"
+	"geniusrabbit.dev/sspserver/internal/endpoint/dynamic"
+	"geniusrabbit.dev/sspserver/internal/endpoint/proxy"
 	"geniusrabbit.dev/sspserver/internal/stream"
-	"geniusrabbit.dev/sspserver/internal/urlgenerator"
-	"geniusrabbit.dev/sspserver/private/templates"
 )
 
 const (
 	eventsStreamName   = "events"
 	userInfoStreamName = "user_info"
 	winStreamName      = "wins"
+	adInfoStreamName   = "ad_info"
 )
 
 var (
-	datasource interface{}
-	config     appcontext.Config
-	commit     = ""
-	appVersion = "develop"
+	config       appcontext.Config
+	buildCommit  = ""
+	buildVersion = "develop"
 )
 
 func init() {
@@ -61,10 +60,11 @@ func init() {
 	fatalError(config.Load(), "config loading")
 
 	// Init new logger object
-	loggerObj, err := newLogger(config.IsDebug(), config.LogLevel, zap.Fields(
-		zap.String("commit", commit),
-		zap.String("version", appVersion),
-	))
+	loggerObj, err := zlogger.New(config.ServiceName, config.LogEncoder,
+		config.LogLevel, config.LogAddr, zap.Fields(
+			zap.String("commit", buildCommit),
+			zap.String("version", buildVersion),
+		))
 	fatalError(err, "configure logger")
 
 	zap.ReplaceGlobals(loggerObj)
@@ -76,269 +76,230 @@ func init() {
 
 func main() {
 	var (
-		sspConf      = &config.AdServer.SSP
-		adServerConf = &config.AdServer
-		storageConf  = &config.AdServer.Storage
 		logger       = zap.L()
+		adServerConf = &config.AdServer
+		adSSPConf    = &config.AdServer.SSP
 		ctx, cancel  = context.WithCancel(context.Background())
 	)
-
 	defer cancel()
 
-	// Init database connection
-	urlData, err := url.Parse(adServerConf.Storage.Connection)
-	fatalError(err, "connection URL parse")
-
-	if urlData.Scheme == "fs" {
-		datasource = adServerConf.Storage.Connection[3:]
-	} else {
-		datasource, err = gorm.Open(urlData.Scheme, adServerConf.Storage.Connection)
-		fatalError(err, "open database")
-	}
+	// Connect to advertisement data source
+	dataAccessor, err := datainit.Connect(ctx, config.AdServer.Storage.Connection)
+	fatalError(err, "advertisement data")
 
 	// Register event streams
 	{
-		// Register events logger
-		eventQueue := adServerConf.EventPipeline.EventQueue
-		{
-			pub, err := stream.ConnectPublisher(ctx, eventQueue.Connection)
-			fatalError(err, "connect to '"+eventQueue.Connection+"' topics")
-			nc.Register(eventsStreamName, pub)
-		}
+		configureEventPipeline(ctx, adServerConf)
 
-		// Register user info loop
-		userQueue := adServerConf.EventPipeline.UserInfoQueue
-		{
-			pub, err := stream.ConnectPublisher(ctx, userQueue.Connection)
-			fatalError(err, "connect to '"+userQueue.Connection+"' topics")
-			nc.Register(userInfoStreamName, pub)
-		}
+		// Run notification listener
+		go nc.Listen(ctx)
 
-		// Register wins info loop
-		winQueue := adServerConf.EventPipeline.WinQueue
-		{
-			pub, err := stream.ConnectPublisher(ctx, winQueue.Connection)
-			fatalError(err, "connect to '"+winQueue.Connection+"' topics")
-			nc.Register(winStreamName, pub)
-		}
+		// Close notification processors
+		defer nc.Close()
 	}
 
-	// Event flow
+	// Event flow processor
 	eventGenerator := eventgenerator.New(config.ServiceName)
 	eventStream := eventstream.New(
 		nc.PublisherByName(eventsStreamName),
 		nc.PublisherByName(userInfoStreamName),
 		eventGenerator,
 	)
+	ctx = eventstream.WithStream(ctx, eventStream)
+
+	// Win processor store into the context of requests
+	winStream := eventstream.WinNotifications(nc.PublisherByName(winStreamName))
+	ctx = eventstream.WithWins(ctx, winStream)
 
 	// URL generator object
-	urlGenerator := &urlgenerator.Generator{
+	urlGenerator := (&urlgenerator.Generator{
 		EventGenerator: eventGenerator,
 		PixelGenerator: pixelgenerator.NewPixelGenerator(adServerConf.TrackerHost),
 		CDNDomain:      adServerConf.CDNDomain,
+		LibDomain:      adServerConf.LibDomain,
 		ClickPattern:   "/click?c={code}",
 		DirectPattern:  "/direct?c={code}",
+		WinPattern:     "/win?c={code}",
+	}).Init()
+
+	// Init side modules
+	datainit.Initialize(config.IsDebug(), urlGenerator)
+
+	// Init format accessor
+	formatAccessor := mustFormatAccessor(ctx, dataAccessor)
+	ctx = formataccessor.WithContext(ctx, formatAccessor)
+
+	// Init company data accessor
+	companyDataAccessor, err := dataAccessor(ctx, "company")
+	fatalError(err, "company data accessor")
+
+	companyAccessor := companyaccessor.NewCompanyAccessor(companyDataAccessor)
+
+	// Init source data accessor
+	sourceDataAccessor, err := dataAccessor(ctx, "source")
+	fatalError(err, "RTB source data accessor")
+
+	// Init advertisement source accessor
+	sourceAccessor, err := adsourceaccessor.NewAccessor(ctx,
+		sourceDataAccessor, companyAccessor, openrtb.NewFactory())
+	fatalError(err, "RTB source accessor")
+
+	// Init target data accessor
+	targetDataAccessor, err := dataAccessor(ctx, "zone")
+	fatalError(err, "target(zone) data accessor")
+
+	targetAccessor := zoneaccessor.NewZoneAccessor(targetDataAccessor, companyAccessor)
+
+	// Configure advertisement source accessor
+	adsourceWrapper, err := adsource.NewMultisourceWrapper(
+		adsource.WithSourceAccessor(sourceAccessor),
+		adsource.WithTimeout(time.Duration(adSSPConf.RequestTimeout)*time.Millisecond),
+		adsource.WithMaxParallelRequests(adSSPConf.MaxParallelRequests),
+	)
+	fatalError(err, "create adsource accessor")
+
+	// HTTP wrapper for extended handlers
+	httpHandlerWrapper := httphandler.NewHTTPHandlerWrapper(
+		nil, // func(*fasthttp.RequestCtx) {},
+		func(*fasthttp.RequestCtx) context.Context { return ctx },
+		logger.With(zap.String("module", "httpserver")),
+	)
+
+	// Init personification client
+	personDetector := personification.Client(&simplepersondetector.SimpleClient{})
+	if config.Person.Connect != "" {
+		personDetector = personification.Connect(http.NewTransport(
+			config.Person.Connect,
+			http.WithTimeout(config.Person.RequestTimeout, config.Person.KeepAliveTimeout),
+		))
 	}
 
-	// Init templates
-	templates.Debug = config.IsDebug()
-	templates.ServiceDomain = ""
-	templates.URLGen = urlGenerator
+	// Init signature
+	signature := personification.Signeture{
+		UUIDName:       config.Person.UUIDCookieName,
+		SessidName:     config.Person.SessiCookiedName,
+		SessidLifetime: config.Person.SessionLifetime,
+		Detector:       personDetector,
+	}
 
-	// Run notification listener
-	go nc.Listen(ctx)
+	// Prepare spy middleware
+	spyMiddleware := middleware.NewSpy(ctx, signature.Whois, signature.SignCookie)
 
-	// Close notification processors
-	defer nc.Close()
-
-	// Create new source accessor which returns preorety iterator to every request
-	sourceAccessor := adsourceaccessor.MustNewAccessor(
-		loaders.SourceReloader(
-			zap.L(),
-			datasourceFrom(datasource, storageConf.Sources),
-			companyGetter,
-			eventStream,
-			nc.PublisherByName(winStreamName),
-		),
-	)
-
-	targetAccessor := targetaccessor.NewSimpleTargetAccessor(
-		loaders.TargetReloader(datasourceFrom(datasource, storageConf.Zones)),
-	)
-	formatAccessor := types.NewSimpleFormatAccessorWithLoader(
-		loaders.FormatLoader(datasourceFrom(datasource, storageConf.Formats)),
-	)
-
-	// Connect to the detector
-	detector, err := newDetector(ctx, config.Person.Connect,
-		config.Person.MaxConn, config.Person.RequestTimeout, config.Person.KeepAliveTimeout)
-	fatalError(err, "Connect detector")
-
-	// Configure advertisement service
-	sspServer, err := ssp.NewServer(
-		ssp.WithBaseSource(nil),
-		ssp.WithSourceAccessor(sourceAccessor),
-		ssp.WithTimeout(time.Duration(sspConf.RequestTimeout)*time.Millisecond),
-		ssp.WithMaxParallelRequests(sspConf.MaxParallelRequests),
-	)
-	fatalError(err, "Create server")
-
+	// HTTP server initialyze
 	server, err := httpserver.NewServer(
 		httpserver.WithDebugMode(config.IsDebug()),
 		httpserver.WithServiceName(config.ServiceName),
-		httpserver.WithZoneAccessor(targetAccessor),
-		httpserver.WithFormatAccessor(formatAccessor),
-		httpserver.WithAdvertisementSource(sspServer),
-		httpserver.WithURLGenerator(urlGenerator),
-		httpserver.WithEventStream(eventStream),
 		httpserver.WithLogger(logger.With(zap.String("module", "httpserver"))),
-		httpserver.WithSpy(newSpy(ctx, &config, detector)),
 		httpserver.WithCustomHTTPServer(&fasthttp.Server{
 			ReadBufferSize: 1 << 20,
-			ReadTimeout:    time.Duration(config.Server.HTTP.ReadTimeout) * time.Millisecond,
-			WriteTimeout:   time.Duration(config.Server.HTTP.WriteTimeout) * time.Millisecond,
+			ReadTimeout:    config.Server.HTTP.ReadTimeout,
+			WriteTimeout:   config.Server.HTTP.WriteTimeout,
 		}),
+		httpserver.WithCustomRouter(func(router *router.Router) {
+			router.ServeFiles("/public/{filepath:*}", "/public")
+		}),
+		httpserver.WithExtensions(
+			pixel.NewExtension(
+				pixel.WithEventStream(eventStream),
+				pixel.WithHTTPHandlerWrapper(httpHandlerWrapper),
+			),
+			trakeraction.NewExtension(
+				trakeraction.WithEventStream(eventStream),
+				trakeraction.WithHTTPHandlerWrapper(httpHandlerWrapper),
+				trakeraction.WithURLGenerator(urlGenerator),
+			),
+			// Register HTTP endpoints extensions
+			endpoint.NewExtension(
+				endpoint.WithAdvertisementSource(adsourceWrapper),
+				endpoint.WithHTTPHandlerWrapper(httpHandlerWrapper),
+				endpoint.WithFormatAccessor(formatAccessor),
+				endpoint.WithZoneAccessor(targetAccessor),
+				endpoint.WithSpy(spyMiddleware),
+				endpoint.WithSendpoints(
+					direct.New(formatAccessor, config.AdServer.Logic.Direct.DefaultURL),
+					dynamic.New(urlGenerator),
+					proxy.New(),
+				),
+			),
+		),
 	)
-	fatalError(err)
-
-	endpoint.Each(func(name string, factory endpoint.Factory) {
-		server.RegisterEdpoint(name, func(source endpoint.Sourcer, options ...interface{}) (endpoint.Endpoint, error) {
-			return factory(source, append(options, urlGenerator, eventStream)...)
-		})
-	})
-
-	// RUN reloader loop
-	timer := reloadLoop(time.Second*30, func() {
-		logError(sourceAccessor.Reload(), "source accessor")
-		logError(targetAccessor.Reload(), "target accessor")
-		logError(formatAccessor.Reload(), "format accessor")
-	})
-
-	onKillApplication(func() {
-		timer.Stop()
-		nc.Close()
-		server.Shutdown()
-	})
+	fatalError(err, "new HTTP server")
 
 	fmt.Println("Run HTTP server", config.Server.HTTP.Listen)
 	fatalError(server.Listen(ctx, config.Server.HTTP.Listen))
 }
 
-func newDetector(ctx context.Context, url string, maxConn int, timeout, keepAlive time.Duration) (cli personification.Client, err error) {
-	switch {
-	case strings.HasPrefix(url, "http://") || strings.HasPrefix(url, "https://"):
-		tr := uhttp.NewTransport(url,
-			uhttp.WithGET,
-			uhttp.WithMaxConn(maxConn),
-			uhttp.WithTimeout(timeout, keepAlive))
-		cli = personification.Connect(tr)
-	case strings.HasPrefix(url, "grpc://"):
-		tr, err := ugrpc.NewTransport(ctx, url, ugrpc.WithSecure(false))
-		if err != nil {
-			return nil, err
+func configureEventPipeline(ctx context.Context, adServerConf *appcontext.AdServerConfig) {
+	// Register events data stream
+	if eventQueue := adServerConf.EventPipeline.EventQueue; eventQueue.Connection != "" {
+		pub, err := stream.ConnectPublisher(ctx, eventQueue.Connection)
+		fatalError(err, "connect to '"+eventQueue.Connection+"' topics")
+		if config.IsDebug() {
+			pub = stream.WrapPublisherWithLog(eventsStreamName, pub)
 		}
-		cli = personification.Connect(tr)
-	default:
-		cli = personification.DummyClient{}
+		nc.Register(eventsStreamName, pub)
+	} else {
+		zap.L().Info("register new events dummy publisher")
+		nc.Register(eventsStreamName, notificationMessageLog("events"))
 	}
-	return cli, err
-}
 
-func newSpy(ctx context.Context, cfg *appcontext.Config, detector personification.Client) middleware.Spy {
-	signature := personification.Signeture{
-		UUIDName:       cfg.Person.UUIDCookieName,
-		SessidName:     cfg.Person.SessiCookiedName,
-		SessidLifetime: cfg.Person.SessionLifetime,
-		Detector:       detector,
-	}
-	return middleware.NewSpy(ctx, signature.Whois, signature.SignCookie)
-}
-
-func reloadLoop(duration time.Duration, fnk func()) *time.Ticker {
-	fnk()
-
-	timer := time.NewTicker(duration)
-	go func() {
-	loop:
-		for {
-			select {
-			case <-timer.C:
-				fnk()
-			case <-nc.OnClose():
-				break loop
-			}
+	// Register user info data stream
+	if userQueue := adServerConf.EventPipeline.UserInfoQueue; userQueue.Connection != "" {
+		pub, err := stream.ConnectPublisher(ctx, userQueue.Connection)
+		fatalError(err, "connect to '"+userQueue.Connection+"' topics")
+		if config.IsDebug() {
+			pub = stream.WrapPublisherWithLog(userInfoStreamName, pub)
 		}
-	}()
-	return timer
+		nc.Register(userInfoStreamName, pub)
+	} else {
+		zap.L().Info("register new userInfo dummy publisher")
+		nc.Register(userInfoStreamName, notificationMessageLog("user_info"))
+	}
+
+	// Register wins info data stream
+	if winQueue := adServerConf.EventPipeline.WinQueue; winQueue.Connection != "" {
+		pub, err := stream.ConnectPublisher(ctx, winQueue.Connection)
+		fatalError(err, "connect to '"+winQueue.Connection+"' topics")
+		if config.IsDebug() {
+			pub = stream.WrapPublisherWithLog(winStreamName, pub)
+		}
+		nc.Register(winStreamName, pub)
+	} else {
+		zap.L().Info("register new wins dummy publisher")
+		nc.Register(winStreamName, notificationMessageLog("win"))
+	}
+
+	// Register adinfo data stream
+	if adInfoQueue := adServerConf.EventPipeline.AdInfoQueue; adInfoQueue.Connection != "" {
+		pub, err := stream.ConnectPublisher(ctx, adInfoQueue.Connection)
+		fatalError(err, "connect to '"+adInfoQueue.Connection+"' topics")
+		nc.Register(adInfoStreamName, pub)
+	} else {
+		zap.L().Info("register new adInfo dummy publisher")
+		nc.Register(adInfoStreamName, notificationMessageLog("adinfo"))
+	}
 }
 
-func onKillApplication(fnk func()) {
-	gracefulStop := make(chan os.Signal)
-	signal.Notify(gracefulStop, syscall.SIGTERM)
-	signal.Notify(gracefulStop, syscall.SIGINT)
-
-	go func() {
-		<-gracefulStop
-		fnk()
-	}()
+func notificationMessageLog(streamName string) nc.FuncPublisher {
+	return func(ctx context.Context, msgs ...any) error {
+		for _, msg := range msgs {
+			zap.L().Debug("notification message",
+				zap.String("stream", streamName),
+				zap.Any("msg", msg),
+			)
+		}
+		return nil
+	}
 }
 
-func fatalError(err error, message ...interface{}) {
+func mustFormatAccessor(ctx context.Context, dataAccessor datainit.DataLoaderAccessorFnk) types.FormatsAccessor {
+	formatDataAccessor, err := dataAccessor(ctx, "format")
+	fatalError(err, "format data accessor")
+	return formataccessor.NewFormatAccessor(formatDataAccessor)
+}
+
+func fatalError(err error, message ...any) {
 	if err != nil {
 		zap.L().Fatal(fmt.Sprint(message...), zap.Error(err))
 	}
-}
-
-func logError(err error, message ...interface{}) error {
-	if err != nil {
-		zap.L().Error(fmt.Sprint(message...), zap.Error(err))
-	}
-	return err
-}
-
-func newLogger(debug bool, loglevel string, options ...zap.Option) (logger *zap.Logger, err error) {
-	if debug {
-		config := zap.NewDevelopmentConfig()
-		config.EncoderConfig.EncodeLevel = zapcore.CapitalColorLevelEncoder
-		return config.Build(options...)
-	}
-
-	var (
-		level         zapcore.Level
-		loggerEncoder = zapcore.NewJSONEncoder(zapcore.EncoderConfig{
-			TimeKey:        "ts",
-			LevelKey:       "level",
-			NameKey:        "logger",
-			CallerKey:      "caller",
-			MessageKey:     "msg",
-			StacktraceKey:  "stacktrace",
-			EncodeLevel:    zapcore.LowercaseLevelEncoder,
-			EncodeTime:     zapcore.ISO8601TimeEncoder,
-			EncodeDuration: zapcore.StringDurationEncoder,
-		})
-	)
-	if err := level.UnmarshalText([]byte(loglevel)); err != nil {
-		logger.Error("parse log level error", zap.Error(err))
-	}
-	core := zapcore.NewCore(loggerEncoder, os.Stdout, level)
-	logger = zap.New(core, options...)
-
-	return logger, nil
-}
-
-// fake companyGetter with unlimited balance
-func companyGetter(id uint64) *models.Company {
-	return &models.Company{
-		ID:       id,
-		Balance:  billing.MoneyFloat(9999999),
-		MaxDaily: billing.MoneyFloat(9999999),
-	}
-}
-
-func datasourceFrom(datasource interface{}, postfix string) interface{} {
-	switch ds := datasource.(type) {
-	case string:
-		return filepath.Join(ds, postfix)
-	}
-	return datasource
 }
