@@ -7,8 +7,8 @@ import (
 	"context"
 	"time"
 
-	nc "github.com/geniusrabbit/notificationcenter"
-	"github.com/geniusrabbit/notificationcenter/nats"
+	nc "github.com/geniusrabbit/notificationcenter/v2"
+	"github.com/geniusrabbit/notificationcenter/v2/nats"
 	natsio "github.com/nats-io/nats.go"
 )
 
